Added sql scripts for auditoriums table and logic

+++ src/main/resources/sql/auditoriums_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE IF NOT EXISTS public.auditoriums
     id SERIAL PRIMARY KEY,
     name text COLLATE pg_catalog."default" NOT NULL,
     number_of_seats integer NOT NULL,
-    vip_seats smallint[] NOT NULL,
+    vip_seats integer[] NOT NULL,
     CONSTRAINT auditoriums_name_key UNIQUE (name)
 );
